backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
scikit-learn
pandas
python-dotenv
sqlalchemy
